requirements.txt
albumentations
catalyst
pytorch
torchvision
tqdm
matplotlib
numpy
scikitlearn
